Modifes class DropUsersTable for possible deployment in render.

+++ db/migrate/20240224072031_drop_users_table.rb
@@ -1,6 +1,6 @@
 class DropUsersTable < ActiveRecord::Migration[7.0]
   def up
-    drop_table :users
+    drop_table :users if ActiveRecord::Base.connection.table_exists? 'users'
   end
 
   def down
